app/requirements.txt
flask
pytz
gunicorn
google-cloud-bigquery
faker
